+++ test/function/samples/deconflict-default-exported-named-function/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'deconflict default exported named function and class expressions'
+};